requirements.txt
# Core dependencies for the Publisher application

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Markdown processing
markdown>=3.5.0

# Reddit API (optional - only install if using Reddit)
# praw>=7.7.0

# Development and testing dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking (optional)
mypy>=1.5.0

# Code formatting (optional)
black>=23.0.0
flake8>=6.0.0

# Documentation (optional)
sphinx>=7.0.0